+++ src/Datagrids/CommentDataGrid.php
@@ -5,6 +5,9 @@ namespace Webbycrown\BlogBagisto\Datagrids;
 use Illuminate\Support\Facades\DB;
 use Webkul\Core\Models\Channel;
 use Webkul\DataGrid\DataGrid;
+use Webbycrown\BlogBagisto\Models\Category;
+use Webbycrown\BlogBagisto\Models\Tag;
+use Webbycrown\BlogBagisto\Models\Blog;
 
 class CommentDataGrid extends DataGrid
 {
@@ -14,12 +17,26 @@ class CommentDataGrid extends DataGrid
 
     public function prepareQueryBuilder()
     {
-        $queryBuilder = DB::table('blog_comments')
-            ->select('blog_comments.id', 'blog_comments.post', 'blog_comments.author', 'blog_comments.email', 'blog_comments.comment', 'blog_comments.status', 'blog_comments.created_at', 'blogs.name as post_name')
-            ->leftJoin('blogs', 'blog_comments.post', '=', 'blogs.id');
+        // $queryBuilder = DB::table('blog_comments')
+        //     ->select('blog_comments.id', 'blog_comments.post', 'blog_comments.author', 'blog_comments.email', 'blog_comments.comment', 'blog_comments.status', 'blog_comments.created_at');
+
+        // return $queryBuilder;
+
+        $loggedIn_user = auth()->guard('admin')->user()->toarray();
+        $user_id = ( array_key_exists('id', $loggedIn_user) ) ? $loggedIn_user['id'] : 0;
+        $role = ( array_key_exists('role', $loggedIn_user) ) ? ( array_key_exists('name', $loggedIn_user['role']) ? $loggedIn_user['role']['name'] : 'Administrator' ) : 'Administrator';
+
+        $queryBuilder = DB::table('blog_comments');
+
+            if ( $role != 'Administrator' ) {
+                $blogs = Blog::where('author_id', $user_id)->get();
+                $post_ids = ( !empty($blogs) && count($blogs) > 0 ) ? $blogs->pluck('id')->toarray() : array();
+                $queryBuilder->whereIn('blog_comments.post', $post_ids);
+            }
+
+            $queryBuilder->select('blog_comments.id', 'blog_comments.post', 'blog_comments.author', 'blog_comments.email', 'blog_comments.comment', 'blog_comments.status', 'blog_comments.created_at');
 
         return $queryBuilder;
-        // $this->setQueryBuilder($queryBuilder);
     }
 
     public function prepareColumns()
@@ -34,12 +51,17 @@ class CommentDataGrid extends DataGrid
         ]);
 
         $this->addColumn([
-            'index'      => 'post_name',
+            'index'      => 'post',
             'label'      => trans('blog::app.datagrid.name'),
             'type'       => 'string',
             'searchable' => true,
             'sortable'   => true,
-            'filterable' => true,
+            'filterable' => false,
+            'closure'    => function ($value) {
+                $post = Blog::where('id', $value->post)->first();
+                $post_name = ( $post && isset($post->name) && !empty($post->name) && !is_null($post->name) ) ? $post->name : '-';
+                return $post_name;
+            },
         ]);
 
         $this->addColumn([
@@ -54,7 +76,7 @@ class CommentDataGrid extends DataGrid
         $this->addColumn([
             'index'      => 'status',
             'label'      => trans('blog::app.datagrid.status'),
-            'type'       => 'string',
+            'type'       => 'boolean',
             'searchable' => true,
             'sortable'   => true,
             'filterable' => true,
@@ -76,41 +98,54 @@ class CommentDataGrid extends DataGrid
             'searchable' => true,
             'sortable'   => true,
             'filterable' => true,
+            'closure' => function ($value) {
+                if ( $value->created_at != '' && $value->created_at != null ) {
+                    return date_format( date_create($value->created_at), 'j F, Y' );
+                } else {
+                    return '-';
+                }
+            },
         ]);
     }
 
     public function prepareActions()
     {
-        $this->addAction([
-            'title' => 'edit',
-            'method' => 'GET',
-            'route' => 'admin.blog.comment.edit',
-            'icon' => 'icon-edit',
-            'url'    => function ($row) {
-                return route('admin.blog.comment.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('blog.comment.edit')) {
+            $this->addAction([
+                'title' => 'edit',
+                'method' => 'GET',
+                'route' => 'admin.blog.comment.edit',
+                'icon' => 'icon-edit',
+                'url'    => function ($row) {
+                    return route('admin.blog.comment.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'title' => 'delete',
-            'method' => 'POST',
-            'route' => 'admin.blog.comment.delete',
-            'icon' => 'icon-delete',
-            'url'    => function ($row) {
-                return route('admin.blog.comment.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('blog.comment.delete')) {
+            $this->addAction([
+                'title' => 'delete',
+                'method' => 'POST',
+                'route' => 'admin.blog.comment.delete',
+                'icon' => 'icon-delete',
+                'url'    => function ($row) {
+                    return route('admin.blog.comment.delete', $row->id);
+                },
+            ]);
+        }
     }
 
     public function prepareMassActions()
     {
-        $this->addMassAction([
-            'type'   => 'delete',
-            'label'  => trans('admin::app.datagrid.delete'),
-            'title'  => 'Delete',
-            'action' => route('admin.blog.comment.massdelete'),
-            'url' => route('admin.blog.comment.massdelete'),
-            'method' => 'POST',
-        ]);
+        if (bouncer()->hasPermission('blog.comment.delete')) {
+            $this->addMassAction([
+                'type'   => 'delete',
+                'label'  => trans('admin::app.datagrid.delete'),
+                'title'  => 'Delete',
+                'action' => route('admin.blog.comment.massdelete'),
+                'url' => route('admin.blog.comment.massdelete'),
+                'method' => 'POST',
+            ]);
+        }
     }
 }
